requirements.txt
# API
fastapi
uvicorn

# AI
langchain
langchain-community
langchain-google-genai
langchain-chroma

# Vector Database
chromadb

# Util
streamlit
pypdf
python-dotenv
tiktoken